move logic to GraphDBSession
