+++ flex/engines/graph_db/database/graph_db_session.h
@@ -80,13 +80,14 @@ class GraphDBSession {
 
   Result<std::vector<char>> Eval(const std::string& input);
 
+#ifdef BUILD_HQPS
   // Evaluate a temporary stored procedure. close the handle of the dynamic lib
   // immediately.
   Result<std::vector<char>> EvalAdhoc(const std::string& input_lib_path);
 
   // Evaluate a stored procedure with input parameters given.
-  Result<std::vector<char>> EvalHqpsProcedure(const std::string& query_name,
-                                              Decoder& decoder);
+  Result<std::vector<char>> EvalHqpsProcedure(const std::string& input);
+#endif
 
   void GetAppInfo(Encoder& result);
