+++ flex/engines/http_server/handler/graph_db_http_handler.cc
@@ -79,7 +79,8 @@ class graph_db_ic_handler : public seastar::httpd::handler_base {
     auto dst_executor = dispatcher_.get_executor_idx();
 
     auto&& content = req->content;
-    // append int32_t value 0 to content
+    // append int32_t value 0 to content, indicate GraphDBSession.Eval() should
+    // be called.
     content.append("\0", 1);
 
     return executor_refs_[dst_executor]
